#365 mv nemo transformation tests into separate dirs
